backend/requirements.txt
pymupdf
flask
flask-cors
dotenv
google-generativeai
